[FIX] stock: Report stock level forecast correction
Should take minimum id instead of max and different
ids for stock moves and quants.

The report was by day and the view by week, which
gave wrong results for the cumulative quantity
when having multiple values in a week (as it was summing them)
We changed the SQL view to be by week also.

Expected dates in the past are all reduced to the current date.